Add MarkupCalculator
Returning to the PriceEstimator disabled tests, the first one fails as it
calculates the wrong number and a small formatting issue. To tackle the
calculation error, a MarkupCalculator is introduced. The degenerate case
of no categories was previously done, so I include a simple test with a single
category.

The MarkupCalculator tests use a mock 'Job' object. The API of the mock will have
to be implemented by the JobParser.