Rename JobParser to Repacking
The JobParser name was no longer informative. In the tests, the mock seemed to
be its own object so there was a disconnect between its name and how it was
used. It sounds more natural for a Repacking to have a base price and
categories associated with it, than some parser.